requirements.txt
alabaster==0.7.10
argon2-cffi==16.3.0
attrs==17.3.0
Babel==2.5.1
bcrypt==3.1.4
certifi==2018.1.18
cffi==1.11.2
chardet==3.0.4
coverage==4.4.2
decorator==4.2.1
defusedxml==0.5.0
Django==2.0.2
django-allauth==0.35.0
django-appconf==1.0.2
django-carton==1.2.1
django-compressor==2.2
django-debug-toolbar==1.9.1
docutils==0.14
flake8==3.5.0
idna==2.6
imagesize==0.7.1
instamojo-wrapper==1.1
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.6
parso==0.1.1
pew==1.1.2
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
pipenv==9.0.3
pkg-resources==0.0.0
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
py==1.5.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
pytest==3.3.1
python3-openid==3.1.0
pytz==2018.3
PyYAML==3.12
rcssmin==1.0.6
requests==2.18.4
requests-oauthlib==0.8.0
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.5
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.4
tox==2.9.1
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
wcwidth==0.1.7
